# backend/openapi.json
{"openapi":"3.1.0","info":{"title":"Research Intelligence Platform","version":"1.0.0"},"paths":{"/papers/":{"get":{"tags":["Papers"],"summary":"Get Papers","operationId":"get_papers_papers__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/PaperResponse"},"type":"array","title":"Response Get Papers Papers  Get"}}}}}},"post":{"tags":["Papers"],"summary":"Create Paper","operationId":"create_paper_papers__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PaperCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PaperResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/papers/{paper_id}":{"get":{"tags":["Papers"],"summary":"Get Paper","operationId":"get_paper_papers__paper_id__get","parameters":[{"name":"paper_id","in":"path","required":true,"schema":{"type":"integer","title":"Paper Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PaperResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/search/arvix":{"get":{"tags":["Search"],"summary":"Search","operationId":"search_search_arvix_get","parameters":[{"name":"query","in":"query","required":true,"schema":{"type":"string","title":"Query"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/search/openalex":{"get":{"tags":["Search"],"summary":"Search Openalex Papers","operationId":"search_openalex_papers_search_openalex_get","parameters":[{"name":"query","in":"query","required":true,"schema":{"type":"string","title":"Query"}},{"name":"max_results","in":"query","required":false,"schema":{"type":"integer","default":5,"title":"Max Results"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/search/ingest":{"post":{"tags":["Search"],"summary":"Ingest Search Result","operationId":"ingest_search_result_search_ingest_post","parameters":[{"name":"source","in":"query","required":true,"schema":{"type":"string","title":"Source"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","additionalProperties":true,"title":"Paper"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/upload/pdf":{"post":{"tags":["Upload"],"summary":"Upload Pdf","operationId":"upload_pdf_upload_pdf_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_upload_pdf_upload_pdf_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/novelty/{paper_id}":{"get":{"tags":["Novelty"],"summary":"Get Novelty","operationId":"get_novelty_novelty__paper_id__get","parameters":[{"name":"paper_id","in":"path","required":true,"schema":{"type":"integer","title":"Paper Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/gaps/{paper_id}":{"post":{"tags":["Research Gaps"],"summary":"Analyze Research Gaps","operationId":"analyze_research_gaps_gaps__paper_id__post","parameters":[{"name":"paper_id","in":"path","required":true,"schema":{"type":"integer","title":"Paper Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/compare/{paper_id_1}/{paper_id_2}":{"get":{"tags":["Compare"],"summary":"Compare Paper Endpoint","operationId":"compare_paper_endpoint_compare__paper_id_1___paper_id_2__get","parameters":[{"name":"paper_id_1","in":"path","required":true,"schema":{"type":"integer","title":"Paper Id 1"}},{"name":"paper_id_2","in":"path","required":true,"schema":{"type":"integer","title":"Paper Id 2"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/chat":{"post":{"tags":["Chat"],"summary":"Chat","operationId":"chat_chat_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/":{"get":{"summary":"Root","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/health":{"get":{"summary":"Health","operationId":"health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"Body_upload_pdf_upload_pdf_post":{"properties":{"file":{"type":"string","contentMediaType":"application/octet-stream","title":"File"}},"type":"object","required":["file"],"title":"Body_upload_pdf_upload_pdf_post"},"ChatRequest":{"properties":{"question":{"type":"string","maxLength":2000,"minLength":1,"title":"Question"},"paper_id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Paper Id"},"top_k":{"type":"integer","maximum":20.0,"minimum":1.0,"title":"Top K","default":8}},"type":"object","required":["question"],"title":"ChatRequest"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"PaperCreate":{"properties":{"title":{"type":"string","title":"Title"},"abstract":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Abstract"},"authors":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Authors"},"doi":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Doi"},"source":{"type":"string","title":"Source"},"source_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Source Id"},"arxiv_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Arxiv Id"},"published_date":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Published Date"},"updated_date":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Updated Date"},"categories":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Categories"},"pdf_url":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Pdf Url"},"pdf_path":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Pdf Path"},"full_text":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Full Text"}},"type":"object","required":["title","source"],"title":"PaperCreate","description":"Schema used when manually creating a paper through the API."},"PaperResponse":{"properties":{"id":{"type":"integer","title":"Id"},"title":{"type":"string","title":"Title"},"abstract":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Abstract"},"authors":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Authors"},"doi":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Doi"},"source":{"type":"string","title":"Source"},"source_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Source Id"},"arxiv_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Arxiv Id"},"published_date":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Published Date"},"updated_date":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Updated Date"},"categories":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Categories"},"pdf_url":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Pdf Url"},"pdf_path":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Pdf Path"},"full_text":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Full Text"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["id","title","source","created_at"],"title":"PaperResponse","description":"Schema returned by the paper API."},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}